update vets website asset scripts, cleanup old cruft
