--- scripts/install-nvm.sh
@@ -1,14 +1,0 @@
-#!/usr/bin/env bash
-
-# Install NVM.
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.3/install.sh | bash
-
-# Make sure we can run it.
-export NVM_DIR="$HOME/.nvm"
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
-[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
-
-# our version managed by nvm
-nvm install 18.18.0
-# version vets-website needs
-nvm install 14.15.1
