--- scripts/install_github_status_updater.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-set -e;
-destination="${1:-/usr/local/bin}";
-version="${2:-0.5.0}";
-executable_path="${destination}/github-status-updater";
-curl --location "https://github.com/cloudposse/github-status-updater/releases/download/${version}/github-status-updater_linux_amd64" --output "${executable_path}";
-chmod +x "${executable_path}";
